+++ R/FullConditionals.R
@@ -68,7 +68,7 @@
 #' @param nn A vector giving the number of observations in each of G groups so that Dirichlet posteriors rather than priors can be sampled from. This defaults to 0, i.e. simulation from the prior. Be warned that this will be recycled if necessary.
 #'
 #' @return A Dirichlet vector of \code{G} weights which sum to 1.
-#' @references Devroye, L. (1986) \emph{Non-Uniform Random Variate Generation}, Springer-Verlag, New York, 1986, p.594.
+#' @references Devroye, L. (1986) \emph{Non-Uniform Random Variate Generation}, Springer-Verlag, New York, p. 594.
 #' @export
 #'
 #' @examples
@@ -263,20 +263,50 @@
       if(any(!is.numeric(shape),
              length(shape) != 1))          stop("'shape' must be a single digit")
       inv.cov   <- try(base::solve(covar), silent=TRUE)
-      if(inherits(inv.cov, "try-error"))  {
-        inv.cov <- .moore_inv(covar)
+      if(inherits(inv.cov, "try-error"))   {
+        covsvd  <- svd(covar)
+        posi    <- covsvd$d > max(sqrt(.Machine$double.eps) * covsvd$d[1L], 0)
+        inv.cov <- if(all(posi)) covsvd$v %*% (t(covsvd$u)/covsvd$d) else if(!any(posi))
+                   array(0, dim(covar)[2L:1L]) else covsvd$v[,posi, drop=FALSE] %*% (t(covsvd$u[,posi, drop=FALSE])/covsvd$d[posi])
       }
         unname((shape - 1)/switch(match.arg(type), unconstrained=diag(inv.cov),
                                   isotropic=rep(exp(mean(log(diag(inv.cov)))), ncol(covar))))
     }
 
   # Alpha/Discount Shifted Gamma Hyperparameters
-    .shift_GA   <- function(shape, rate, shift = 0L, param = c("rate", "scale")) {
-      var       <- shape/rate^2
-      exp       <- var  * rate + shift
+#' Moment Matching Parameters of Shifted Gamma Distributions
+#'
+#' This function takes shape and rate parameters of a Gamma distribution and modifies them to achieve the same expected value and variance when the left extent of the support of the distribution is shifted up or down.
+#' @param shape Shape parameter a of a Gamma(a, b) distribution. Must be strictly positive.
+#' @param rate Rate parameter b of a Gamma(a, b) distribution. Must be strictly positive.
+#' @param shift Modifier, such that the Gamma distribution has support on (\code{shift}, \eqn{\infty}). Can be positive or negative, though typically negative and small.
+#' @param param Switch controlling whether the supplied \code{rate} parameter is indeed a rate, or actually a scale parameter. Also governs whether the output is given in terms of rate or scale. Defaults to "\code{rate}".
+#'
+#' @return A list of length 2, containing the modified shape and rate parameters, respectively.
+#' @export
+#'
+#' @author Keefe Murphy
+#'
+#' @examples
+#' # Shift a Ga(shape=4, rate=2) distribution to the left by 1;
+#' # achieving the same expected value of 2 and variance of 1.
+#' shift_GA(4, 2, -1)
+    shift_GA    <- function(shape, rate, shift = 0L, param = c("rate", "scale")) {
+      if(length(shape) > 1 ||
+        !is.numeric(shape) || shape <= 0) stop("Argument 'shape' must be a single strictly positive number")
+      if(length(rate)  > 1 ||
+        !is.numeric(rate)  || rate  <= 0) stop("Argument 'rate' must be a single strictly positive number")
+      if(length(shift) > 1 ||
+        !is.numeric(shift))               stop("Argument 'shift' must be a single number")
+      param     <- match.arg(param)
+      rate      <- switch(param, rate=rate, 1/rate)
+      exp       <- shape/rate
+      if(shift  >= exp)                   warning("This expected value is not achievable with the supplied 'shift'", call.=FALSE)
+      var       <- exp/rate
+      exp       <- pmax(var * rate   - shift, 0)
       rate      <- exp/var
       shape     <- rate * exp
-        return(list(shape = shape, rate = switch(match.arg(param), rate=rate, 1/rate)))
+        return(list(shape   = shape, rate = switch(param, rate=rate, 1/rate)))
     }
 
   # Check Shrinkage Hyperparemeters
@@ -479,33 +509,134 @@
     }
 
   # Positive-(Semi)Definite Checker
-
     #' Check Postive-(Semi)definiteness of a matrix
     #'
-    #' Tests whether all eigenvalues of a symmetric matrix are positive (or strictly non-negative).
+    #' Tests whether all eigenvalues of a symmetric matrix are positive (or strictly non-negative) to check for positive-definiteness and positive-semidefiniteness, respectively. If the supplied matrix doesn't satisfy the test, the nearest matrix which does can optionally be returned.
     #' @param x A matrix, assumed to be real and symmetric.
     #' @param tol Tolerance for singular values and for absolute eigenvalues - only those with values larger than tol are considered non-zero (default: tol = \code{max(dim(x))*max(E)*.Machine$double.eps}, where \code{E} is the vector of absolute eigenvalues).
     #' @param semi Logical switch to test for positive-semidefiniteness when \code{TRUE} or positive-definiteness when \code{FALSE} (the default).
+    #' @param make Logical switch to return the nearest matrix which satisifies the test - if the test has been passed, this is of course just \code{x} itself, otherwise the nearest positive-(semi)definite matrix. Note that for reasons due to finite precision arithmetic, finding the nearest positive-definite and nearest positive-semidefinite matrices are effectively equivalent tasks.
     #'
-    #' @return A logical value (\code{TRUE} or \code{FALSE})
+    #' @return If \code{isTRUE(make)}, a list with two components:
+    #' \describe{
+    #' \item{check}{A logical value indicating whether the matrix satisfies the test.}
+    #' \item{X.new}{The nearest matrix which satisfies the test (which may just be the input matrix itself.)}
+    #' }
+    #' Otherwise, only the logical value indicating whether the matrix satisfies the test is returned.
+    #'
     #' @export
     #'
     #' @examples
-    #' x <- cov(matrix(rnorm(100), nrow=10, ncol=10))
+    #' x    <- cov(matrix(rnorm(100), nrow=10, ncol=10))
     #' is.posi_def(x)
     #' is.posi_def(x, semi=TRUE)
-    is.posi_def <- function(x, tol = NULL,   semi = FALSE)  {
-      if(!is.matrix(x)   &&
-         nrow(x) != ncol(x))               stop("argument x is not a square matrix")
+    #'
+    #' Xnew <- is.posi_def(x, semi=FALSE, make=TRUE)$X.new
+    #' identical(x, Xnew)
+    #' identical(x, is.posi_def(x, semi=TRUE, make=TRUE)$X.new)
+    is.posi_def <- function(x, tol = NULL, semi = FALSE, make = FALSE)  {
+      if(!is.matrix(x)     &&
+        nrow(x) != ncol(x))                stop("argument x is not a square matrix")
       if(!is.symmetric(x))                 stop("argument x is not a symmetric matrix")
-      if(!is.numeric(x))                   stop("argument x is not a numeric matrix")
-      eigs      <- eigen(x,  only.values = TRUE,  symmetric = TRUE)$values
-      abseigs   <- abs(eigs)
-      tol       <- if(missing(tol)) max(abseigs)  * nrow(x) * .Machine$double.eps else tol
-      if(length(tol) > 1 ||
+      if(!is.double(x))                    stop("argument x is not a numeric matrix")
+      if(!is.logical(semi) ||
+         length(semi) > 1)                 stop("argument semi is not a single logical indicator")
+      if(!is.logical(make) ||
+         length(make) > 1)                 stop("argument make is not a single logical indicator")
+      d         <- nrow(x)
+      eigs      <- eigen(x, symmetric = TRUE)
+      eval      <- eigs$values
+      abseigs   <- abs(eval)
+      tol       <- if(missing(tol)) max(abseigs) * d * .Machine$double.eps else tol
+      if(length(tol)  > 1  ||
          !is.numeric(tol))                 stop("argument tol is not a single number")
-      test      <- replace(eigs, abseigs < tol, 0)
-        !any(if(isTRUE(semi)) test < 0 else test <= 0)
+      test      <- replace(eval, abseigs < tol, 0)
+      check     <- !any(if(isTRUE(semi)) test < 0 else test <= 0)
+      if(isTRUE(make))  {
+        evec    <- eigs$vectors
+        return(list(check = check, X.new = if(all(check)) x else x + evec %*% tcrossprod(diag(pmax(ifelse(isTRUE(semi), 0, .Machine$double.eps), 2 * tol - eval), d), evec)))
+      } else check
+    }
+
+  # Ledermann Bound
+#' Ledermann Bound
+#'
+#' Returns the maximum possible number of latent factors in a factor analysis model for data of dimension \code{P}. This Ledermann bound is given by the largest integer smaller than or equal to the solution \eqn{k}{k} of \eqn{(M - k)^2 \geq M + k}{(M - k)^2 >= M + k}.
+#' @param P Integer number of variables in data set.
+#'
+#' @return The Ledermann bound, a non-negative integer.
+#' @export
+#'
+#' @examples
+#' Ledermann(25)
+    Ledermann   <- function(P) {
+      P         <- as.integer(P)
+      if(length(P)   > 1  || P <= 0)       stop('argument P is a not a single positive integer')
+      R         <- P + 0.5 * (1 - sqrt(8 * P  + 1))
+        as.integer(floor(ifelse(1e-10 > abs(R - round(R)), round(R), R)))
+    }
+
+  # Procrustes Transformation
+#' Procrustes Transformation
+#'
+#' This function performs a Procrustes transformation on a matrix \code{X} to minimize the squared distance between \code{X} and another comparable matrix \code{Xstar}.
+#' @param X The matrix to be transformed.
+#' @param Xstar The target matrix.
+#' @param translate Logical value indicating whether \code{X} should be translated (defaults to \code{FALSE}).
+#' @param dilate Logical value indicating whether \code{X} should be dilated (defaults to \code{FALSE}).
+#' @param sumsq Logical value indicating whether the sum of squared differences between \code{X} and \code{Xstar} should be calculated and returned.
+#'
+#' @details{
+#'    \code{R}, \code{tt}, and \code{d} are chosen so that:
+#'
+#'    \deqn{d \times \mathbf{X} \mathbf{R} + 1\hspace*{-3pt}1 \underline{t}^\top \approx X^\star}{d X R + 1 t' approximately Xstar}
+#'
+#'    \code{X.new} is given by:
+#'
+#'    \deqn{X_{\textrm{new}} = d \times \mathbf{X} \mathbf{R} + 1\hspace*{-3pt}1 \underline{t}^\top}{X.new = d X R + 1 t'}
+#'}
+#'
+#' @return A list containing:
+#' \describe{
+#' \item{X.new}{The matrix that is the Procrustes transformed version of \code{X}.}
+#' \item{R}{The rotation matrix.}
+#' \item{t}{The translation vector (if \code{isTRUE(translate)}).}
+#' \item{d}{The scaling factor (is \code{isTRUE(dilate)}).}
+#' \item{ss}{The sum of squared differences (if \code{isTRUE(sumsq)}).}
+#' }
+#' @export
+#'
+#' @references Borg, I. and Groenen, P. J. F. (1997) \emph{Modern Multidimensional Scaling}. Springer-Verlag, New York, pp. 340-342.
+#'
+#' @examples
+#' # Match two matrices, allowing translation and dilation
+#' mat1    <- diag(rnorm(10))
+#' mat2    <- 0.05 * matrix(rnorm(100), 10, 10) + mat1
+#' proc    <- Procrustes(X=mat1, Xstar=mat2, translate=TRUE, dilate=TRUE, sumsq=TRUE)
+#'
+#' # Extract the transformed matrix, rotation matrix, translation vector and scaling factor
+#' mat_new <- proc$X.new
+#' mat_rot <- proc$R
+#' mat_t   <- proc$t
+#' mat_d   <- proc$d
+#'
+#' # Compare the sum of squared differences to a Procestean transformation with rotation only
+#' mat_ss  <- proc$ss
+#' mat_ss2 <- Procrustes(X=mat1, Xstar=mat2, sumsq=TRUE)$ss
+    Procrustes  <- function(X, Xstar, translate = FALSE, dilate = FALSE, sumsq = FALSE) {
+      N         <- nrow(X)
+      P         <- ncol(X)
+      if(N      != nrow(Xstar))            stop("X and Xstar do not have the same number of rows")
+      if(P      != ncol(Xstar))            stop("X and Xstar do not have the same number of columns")
+      J         <- if(translate) diag(N) - matrix(1/N, N, N)                         else diag(N)
+      C         <- crossprod(Xstar, J) %*% X
+      svdX      <- svd(C)
+      R         <- tcrossprod(svdX$v, svdX$u)
+      d         <- if(dilate)    sum(diag(C %*% R))/sum(diag(crossprod(X, J) %*% X)) else 1
+      tt        <- if(translate) crossprod(Xstar - d * X %*% R, matrix(1, N, 1))/N   else 0
+      X.new     <- d * X %*% R + if(translate) matrix(tt, N, P, byrow = TRUE)        else tt
+        return(c(list(X.new = X.new), list(R = R), if(translate) list(t = tt),
+                 if(dilate) list(d = d), if(sumsq) list(ss = sum((X - X.new)^2))))
     }
 
   # Length Checker
@@ -719,7 +850,14 @@
     }
 
   # Other Hidden Functions
-    .chol       <- function(x) tryCatch(chol(x), error=function(e) chol(.make_posdef(x)))
+    .chol       <- function(x) tryCatch(chol(x), error=function(e) {
+      d         <- nrow(x)
+      eigs      <- eigen(x, symmetric = TRUE)
+      eval      <- eigs$values
+      evec      <- eigs$vectors
+        return(chol(x + evec %*% tcrossprod(diag(pmax(0,  2 * max(abs(eval)) * d * .Machine$double.eps - eval), d), evec)))
+      }
+    )
 
     .detach_pkg <- function(pkg, character.only = FALSE) {
       if(!character.only) {
@@ -738,13 +876,15 @@
         if(ent  %in% c("exit", "EXIT"))    stop()
     }
 
-    .ledermann  <- function(N, P) {
-      R         <- P + 0.5 - (0.5  * sqrt(8 * P + 1))
-        as.integer(floor(min(N - 1, ifelse(1e-10 > abs(R - round(R)), round(R), R))))
+    .logdensity     <- function(x, left = 0)  {
+      d        <- density(x,     bw   = "SJ")
+      h        <- d$bw
+      w        <- 1/pnorm(left,  mean = x, sd = h, lower.tail = FALSE)
+        return(suppressWarnings(density(x, bw = h, kernel = "gaussian", weights = w/length(x))))
     }
 
     .logitdensity   <- function(x)  {
-      y         <- qlogis(x)
+      y         <- qlogis(x[x  > 0  & x < 1])
       g         <- density(y, bw    = "SJ")
       xgrid     <- plogis(g$x)
       g$y       <- g$y/(xgrid  * (1 - xgrid))
@@ -752,52 +892,7 @@
         return(g)
     }
 
-   .make_posdef <- function(x, tol = NULL) {
-      if(!is.matrix(x)   &&
-        nrow(x) != ncol(x))                stop("argument x is not a square matrix")
-      if(!is.symmetric(x))                 stop("argument x is not a symmetric matrix")
-      if(!is.numeric(x))                   stop("argument x is not a numeric matrix")
-      d         <- nrow(x)
-      eigs      <- eigen(x, symmetric = TRUE)
-      eval      <- eigs$values
-      evec      <- eigs$vectors
-      tol       <- if(missing(tol)) max(abs(eval)) * d * .Machine$double.eps else tol
-        return(x + evec %*% tcrossprod(diag(pmax(0,  2 * tol - eval), d), evec))
-    }
-
-    .moore_inv  <- function(X, tol = sqrt(.Machine$double.eps)) {
-      if(length(dim(X)) > 2L ||
-         !(is.numeric(X)     ||
-           is.complex(X)))                 stop("'X' must be a numeric or complex matrix")
-      if(!is.matrix(X))    X <- as.matrix(X)
-      Xs        <- svd(X)
-      if(is.complex(X)) Xs$u <- Conj(Xs$u)
-      Posi      <- Xs$d > max(tol * Xs$d[1L], 0)
-      if(all(Posi))     {
-          Xs$v %*% (t(Xs$u)/Xs$d)
-      } else if(!any(Posi))   {
-          array(0, dim(X)[2L:1L])
-      } else {
-          Xs$v[,Posi,drop=FALSE] %*% (t(Xs$u[,Posi,drop=FALSE])/Xs$d[Posi])
-      }
-    }
-
     .power2     <- function(x) x * x
-
-    .Procrustes <- function(X, Xstar, translate = FALSE, dilate = FALSE) {
-      n         <- nrow(X)
-      p         <- ncol(X)
-      if(n      != nrow(Xstar))            stop("X and Xstar do not have the same number of rows")
-      if(p      != ncol(Xstar))            stop("X and Xstar do not have the same number of columns")
-      J         <- if(translate) diag(n) - matrix(1/n, n, n)                         else diag(n)
-      C         <- crossprod(Xstar, J) %*% X
-      svdX      <- svd(C)
-      R         <- tcrossprod(svdX$v, svdX$u)
-      s         <- if(dilate)    sum(diag(C %*% R))/sum(diag(crossprod(X, J) %*% X)) else 1
-      tt        <- if(translate) crossprod(Xstar - s * X %*% R, matrix(1, n, 1))/n   else 0
-      Xnew      <- s * X %*% R + if(translate) matrix(tt, n, p, byrow = TRUE)        else tt
-        return(c(list(X.new = Xnew), list(R = R), if(translate) list(tt = tt), if(dilate) list(s = s)))
-    }
 
     .which0     <- function(x) which(x == 0)
     #