+++ R/SimulateData.R
@@ -37,7 +37,7 @@ sim_IMIFA_data <- function(N = 300L, G = 3L, P = 50L, Q = rep(4L, G), pis = rep(
          length(G) != 1, length(P) != 1)) stop("'N', 'P', 'G', and 'loc.diff' must be of length 1")
   if(!is.numeric(loc.diff))               stop("'loc.diff' must be numeric")
   if(any(N  < 2, N <= G))                 stop("Must simulate more than one data-point and the number of groups cannot exceed N")
-  if(any(Q  > .ledermann(N=N, P=P)))      warning(paste0("Are you sure you want to generate this many factors relative to N=", N, " and P=", P, "?"), call.=FALSE)
+  if(any(Q  > min(N - 1, Ledermann(P))))  warning(paste0("Are you sure you want to generate this many factors relative to N=", N, " and P=", P, "?"), call.=FALSE)
   if(length(Q) != G) {
     if(!missing(Q))  {
       if(length(Q) == 1) {
@@ -92,9 +92,7 @@ sim_IMIFA_data <- function(N = 300L, G = 3L, P = 50L, Q = rep(4L, G), pis = rep(
     covmat     <- provideDimnames(diag(psi.true) + switch(method, marginal=tcrossprod(l.true), 0), base=list(vnames))
     if(!all(is.symmetric(covmat),
             is.double(covmat)))           stop("Invalid covariance matrix")
-    if(!is.posi_def(covmat)) {
-      covmat   <- .make_posdef(covmat)
-    }
+    covmat     <- is.posi_def(covmat, make=TRUE)$X.new
     sigma      <- if(any(Q.g > 0, method == "conditional")) .chol(covmat) else sqrt(covmat)
     means      <- matrix(mu.true, nrow=N.g, ncol=P, byrow=TRUE) + switch(method, conditional=tcrossprod(eta.true[true.zlab == g, seq_len(Q.g), drop=FALSE], l.true), 0)
     simdata    <- rbind(simdata, means + matrix(rnorm(N.g * P), nrow=N.g, ncol=P) %*% sigma)